backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.7
pandas==2.2.3
scikit-learn==1.3.2
openpyxl==3.1.2
python-dotenv==1.0.0
flask-cors==4.0.0
flask-jwt-extended==4.4.4
requests==2.31.0
gunicorn==20.1.0
python-socketio==5.9.0
